video: temporary fix SkipSeconds
There has to be done some more investigation. Need to lock the threads against each other.
The problem is, that VideoDecodeInput/VideoRenderFrame places a frame in FramesRb after FlushDisplayThread
is done. Setting render->Flushing is not implemented threadsafe.
This happens somewhere around the usleep and goto. Minimizing the usleep timer makes it work better.

But this is still a TODO!

Signed-off-by: Andreas Baierl <ichgeh@imkreisrum.de>