fix trickspeed
This finally should fix trickspeed and stillpicture.

backward trickspeed:
We need to force draining the decoder by sending NULL with avcodec_send_packet when
doing backward trickspeed, because the decoder can decide on it's, how many decoded
frames it wants to buffer before sending the first frame. Normally, the decoder resorts
the frames by ascending PTS, which leads to wrong results in backward trickspeed.
We can trust VDR, that only I-Frames are sent in backward trickspeed, so they can be
decoded without any reference frames. Though it's safe to force receiving a frame after
sending one packet.

After forcing draining we need to flush the decoder in order to be able to decode
new packets. This should be done with avcodec_flush_buffers.
There is some bug in the ffmpeg decoder though - avcodec_flush_buffers doesn't work
when we sent the end-of-stream. So we need to close and reopen the decoder.

We keep this workaround for now, because it's working with other decoders, too.
As soon as avcodec_flush_buffers works on rpi's ffmpeg, use that one, because it's
less expensive.

In order to be able to decide, how many pkts are necessary to decode a frame, we set
trickpkts when starting a new stream. 576i(mpeg2) and 720p(h264) need 1 pkt, 1080i(h264)
need 2 pkts. hevc isn't tested yet...

forward trickspeed:
In forward trickspeed, we don't force the draining, because slow forward is just
like normal playback with all frames but slower and in fast forward mode VDR initiates
the necessary clears for us.

Play/SetPlayMode:
We always reset the decoder when we start playing a new file or channel and we also
reset it when entering normal playback after a backward trickspeed mode.

More changes:
- refactor close stream
- fix stillpicture for hwdec interlaced streams
- fix ffmpeg logging

Tested on RK3399 and RPI4.

Signed-off-by: Andreas Baierl <ichgeh@imkreisrum.de>