fix amlogic h264
playback works again, but sth is wrong with decoder in trickspeed -
needs investigation, if this is a softhddevice or decoder issue

Amlogic is the only known decoder atm, which needs a width and height at decoder start.
Rockchip and RPI do not need this.

Introduce some flags depending on the hardware.
Also introduce a flag which indicates, if the codec works properly when flushing the codec
after a forced drain.

Signed-off-by: Andreas Baierl <ichgeh@imkreisrum.de>