fix segv while ClearVideo()
With the codec reopen workaround used in ClearVideo() we run into a situation,
where VideoCtx gets deleted. Decoding - which calls VideoCtx - is running in
another thread, which can result in a SEGV when VideoCtx-> is wanted.

Introduce a StreamWait variable, which synchronizes the two threads.
We now wait with ClearVideo() until VideoDecodInput has finished the last
frame and it now waits for the ClearVideo() until it continues.

This is not the most beautiful solution and can be improved for sure...

Signed-off-by: Andreas Baierl <ichgeh@imkreisrum.de>