video: try to fix timing issues
In case we have osd but no video, the video was skipped and osd was displayed.
If we have no frames available, which can be displayed, we now wait up to 25ms for a frame, which
is more than the time between 2 pageflips on 50Hz. After that time period the osd is displayed because
we are most likely in a channel switch.

If the decoder can't produce a frame with a vsync period we have a problem anyway.

There was a problem reported with 4K output, replaying recorded UHD content and opening
the VDR replay menu. This is patch is just a guess and needs more testing.

In case it doesn't solve the problem, it should stay nevertheless because it shouldn't hurt anything.

Signed-off-by: Andreas Baierl <ichgeh@imkreisrum.de>