opengl: fix subtitles/osd conflict
OpenGL always did a clear if any Osd should be deleted.
If normal Osd (Level < 10) was displayed and we also had subtitles
on Level 10, the subtitle Osd got frequently deleted, but this
ended up in a clear for the active Osd.
Now we only "clear" active Osds.

Some cleanup in the code was done, too.

Signed-off-by: Andreas Baierl <ichgeh@imkreisrum.de>