osd: fix segfault with software rendering
SkinElchiHD creates buffers, that are bigger than the osd buffer to be rendered in.
Softhddevice clips the buffer to the osd size, but the "bigger" source stride results
in a segfault when doing the memcpy into the osd buffer which has a smaller stride.
We cannot use the destination stride every time, because sometimes we want to render
a smaller area into the osd, which results in a segfault at the sources memory.

Solution is to use the minimum stride.

For now, the segfault has only happened with SkinElchiHD, so imho the bigger width (+1)
should be fixed there.

Signed-off-by: Andreas Baierl <ichgeh@imkreisrum.de>