fix interlacing
If some stream does not provide the framerate and a progressive frame arrives
within an interlaced stream, the frame goes directly to the rendering
framebuffer, which results in a deadlock.

Fix this by setting the interlaced switch depending on an active deinterlacing
filter if this situation occurs.

Maybe there could still be a problem, if the first frame in an interlaced stream
is progressive.