fix trickspeed
Fix the Play()->Pause()->Forward case

Now we wait in Pause() for frames in the filter ringbuffer and then close the filter.
VDR doesn't send a DeviceClear() before slow forward, so in interlaced streams we
have an active filter thread and send non-interlaced frames in trickspeed.

This blocks rendering, because the filter ringbuffer doesn't get empty anymore.
We can't do a Clear() here, because that drops the frames, which have already been
sent by VDR.