version 1.0.0
Bump the version to a new major number,
because the code rewrite is finished for now.

There is really much more to do, but this can be done later.
Focus should be on testing this version and fix bugs - if there are,
then address the remaining issues.

More code improvements could by done on-the-fly or step-by-step.

See all the changes in the commit history. Most of the changes
have been split into separate commits, but sometimes the changes
are squashed into one commit. Not all of the commits build and run,
so bisecting a problem may be difficult.

There have also been some functional changes and improvements
(especially the Play()/Pause()/TrickSpeed() thing) in between
the refactoring commits.

The refactoring commits should have no functional changes the most
time, but that can not be guaranteed. If so, I was too lazy for a rebase ...

Signed-off-by: Andreas Baierl <ichgeh@imkreisrum.de>